requirements.txt
streamlit>=1.28.0
google-cloud-firestore>=2.11.0
google-auth>=2.23.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
plotly>=5.15.0
pandas>=1.5.0
uuid
